requirements.txt
git+https://github.com/Rapptz/discord.py.git
aiomysql
pyyaml
pynacl
orjson
discord-ext-fslash
jishaku
yt-dlp